% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bshrink_stats.R
\name{bshrink_stats}
\alias{bshrink_stats}
\title{Calculate Performance Statistics for Bayesian Shrinkage Models}
\usage{
bshrink_stats(B0, PostB)
}
\arguments{
\item{B0}{Dataframe. A dataframe containing the true regression coefficients (\code{value}) and their corresponding \code{metric} (beta names).}

\item{PostB}{Dataframe. A summary of posterior regression coefficients from the model, including posterior means (\code{mean}), lower HDI bounds (\code{hdilower}), and upper HDI bounds (\code{hdiupper}).}
}
\value{
A list containing the following performance statistics:
\itemize{
\item \code{MSE}: Mean Squared Error.
\item \code{bias}: The average bias of the estimates.
\item \code{CP}: Coverage probability, the proportion of true values falling within the HDI intervals.
\item \code{MIW}: Mean interval width, the average width of the HDI intervals.
\item \code{sens}: Sensitivity, the proportion of true signals correctly identified as non-zero.
\item \code{spec}: Specificity, the proportion of non-signals correctly identified as zero.
}
}
\description{
This function calculates various performance metrics for a Bayesian shrinkage model, including
Mean Squared Error (MSE), bias, coverage probability, mean interval width (MIW), sensitivity, and specificity.
It compares the true regression coefficients with the posterior summary of regression coefficients and
computes these statistics based on the Highest Posterior Density (HPD) intervals.
}
\details{
The function removes iterations and irrelevant simulations from the posterior data (\code{PostB}), and joins the posterior estimates (mean and HDI intervals) with the true coefficients (\code{B0}).
It then calculates various statistics such as bias, MSE, sensitivity, specificity, coverage probability, and mean interval width based on the comparison between the true values and the posterior estimates.
}
\examples{
# Example usage:
B0 <- data.frame(metric = c("beta1", "beta2", "beta3"), value = c(0.5, 0.7, 0))
PostB <- data.frame(metric = rep(c("beta1", "beta2", "beta3"), each = 3),
                    statistic = rep(c("mean", "hdilower", "hdiupper"), times = 3),
                    value = rnorm(9))
results <- bshrink_stats(B0, PostB)
print(results)

}
