% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_y.R
\name{generate_y}
\alias{generate_y}
\title{Generate Binary Response Variable}
\usage{
generate_y(df, beta)
}
\arguments{
\item{df}{Dataframe or matrix. The dataset of predictor variables for which the response variable is generated.}

\item{beta}{Numeric vector. The vector of coefficients corresponding to the predictor variables.}
}
\value{
A binary response vector generated from a Bernoulli distribution.
}
\description{
This function generates a binary response variable \(y\) for a given dataset and
a vector of coefficients. The function calculates the log-odds for each observation,
applies the inverse logit function to obtain probabilities, and then generates a
binary response using a Bernoulli distribution based on these probabilities.
}
\details{
The function first converts the input dataset and coefficient vector into matrices. It then
computes the log-odds by multiplying the predictor matrix by the coefficient matrix. The log-odds
are transformed into probabilities using the inverse logit function. Finally, a binary response is
generated by drawing from a Bernoulli distribution with the computed probabilities.
}
\examples{
# Example usage
df <- matrix(rnorm(100), nrow = 10, ncol = 10)
beta <- rnorm(10)
y <- generate_y(df, beta)

}
