% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierBayes.R
\name{hierBayes}
\alias{hierBayes}
\title{Hierarchical Bayesian Shrinkage with Horseshoe Prior}
\usage{
hierBayes(data, main, demo, MCMC)
}
\arguments{
\item{data}{Dataframe. The full dataset including the response and predictors.}

\item{main}{Integer. The number of main effects in the model.}

\item{demo}{Integer. The number of demographic effects (covariates) in the model.}

\item{MCMC}{Integer. The number of iterations to run MCMC.}
}
\value{
A list containing:
\itemize{
\item \code{BETA}: A matrix of the posterior samples of regression coefficients (main and interaction effects).
\item \code{ETA}: A matrix of the local shrinkage parameters for main effects.
\item \code{THETA}: A matrix of the local shrinkage parameters for interaction effects.
\item \code{A}: A vector of the global shrinkage parameters for main effects.
\item \code{B}: A vector of the global shrinkage parameters for interaction effects.
}
}
\description{
This function implements a hierarchical Bayesian shrinkage method with a horseshoe prior for both main and interaction effects.
It uses Markov Chain Monte Carlo (MCMC) to sample from the posterior distribution and updates the global and local shrinkage
parameters iteratively using the horseshoe prior.
}
\details{
This function initializes the necessary parameters for Bayesian shrinkage using the horseshoe prior, including latent variables, precision parameters, and shrinkage parameters.
The horseshoe prior introduces auxiliary parameters (\code{nu}, \code{delta}, \code{kappa}, and \code{alpha}) to allow for heavy-tailed shrinkage that is particularly effective for sparse models.
It then iteratively updates the coefficients, global, and local shrinkage parameters for both main and interaction effects using MCMC.
}
\note{
This function is adapted from the work found at \url{https://github.com/debamitakundu8/Bayesian-Inference-of-Chemical-Mixtures-in-Risk-Assessment-Incorporating-the-Hierarchical-Principle}.
}
\examples{
# Example usage:
data <- matrix(rnorm(1000), nrow = 100, ncol = 10)
result <- hierBayes(data, main = 5, demo = 2, MCMC = 1000)
print(result$BETA)

}
