% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_shrinkage.R
\name{shared_shrinkage}
\alias{shared_shrinkage}
\title{Hierarchical Shrinkage Model with MCMC for Main and Interaction Effects}
\usage{
shared_shrinkage(data, main, demo, MCMC, hier = "Y")
}
\arguments{
\item{data}{Dataframe or matrix. The full dataset including the response variable and predictors.}

\item{main}{Integer. The number of main effects (predictors) in the model.}

\item{demo}{Integer. The number of demographic effects in the model.}

\item{MCMC}{Integer. The number of iterations to run the MCMC algorithm.}

\item{hier}{Character. Specifies whether to use a hierarchical model for interaction effects ("Y" or "N").}
}
\value{
A list containing the posterior samples of the main effects coefficients (BETA),
local shrinkage parameters for main effects (ETA), local shrinkage parameters for interaction
effects (THETA), and the global shrinkage parameters for main and interaction effects (A and B).
}
\description{
This function fits a hierarchical Bayesian shrinkage model to the provided dataset,
which includes both main effects and two-way interaction effects. It utilizes Markov
Chain Monte Carlo (MCMC) to estimate the coefficients, shrinkage parameters, and precision terms
for both main and interaction effects. The user can specify whether to include hierarchical
structure for interaction effects.
}
\details{
The function performs MCMC sampling to estimate the coefficients of both main effects and
interaction terms, along with the global and local shrinkage parameters. It allows for the
inclusion of a hierarchical structure for the interaction effects, where the shrinkage for
interaction terms depends on the shrinkage of the main effects. The precision parameters for
the latent variables and the coefficients are updated iteratively using Gibbs sampling.
The global and local shrinkage parameters are updated using Gamma distributions.
}
\note{
This function is adapted from the work found at \url{https://github.com/debamitakundu8/Bayesian-Inference-of-Chemical-Mixtures-in-Risk-Assessment-Incorporating-the-Hierarchical-Principle}.
}
\examples{
# Example usage:
data <- matrix(rnorm(1000), nrow = 100, ncol = 10)
results <- shared_shrinkage(data, main = 5, demo = 2, MCMC = 1000, hier = "Y")

}
