% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_with_hdi.R
\name{summarize_with_hdi}
\alias{summarize_with_hdi}
\title{Summarize Data with 95\% HPD (Highest Posterior Density) Intervals}
\usage{
summarize_with_hdi(data, group, num = 0.95)
}
\arguments{
\item{data}{Dataframe. The dataset to be summarized, containing the variables of interest.}

\item{group}{Character. The name of the grouping variable to group the data by.}

\item{num}{Numeric. The confidence interval for the HPD calculation. Defaults to 0.95.}
}
\value{
A long-format dataframe with summary statistics and HPD intervals for each variable, organized by group.
}
\description{
This function calculates summary statistics by groups, including the 95\% Highest Posterior Density (HPD) intervals.
It computes the minimum, mean, standard deviation, maximum, and HPD lower and upper bounds for each group.
}
\details{
The function groups the data by the specified \code{group} variable and calculates the following summary statistics
for each variable: minimum, mean, standard deviation, maximum, and the lower and upper bounds of the HPD interval.
The results are then pivoted into a long format, and the variable names are split into two separate columns (metric and statistic).
}
\examples{
# Example usage:
data <- data.frame(sim = rep(1:5, each = 10), value = rnorm(50))
summarize_with_hdi(data, "sim", 0.95)

}
