% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_function.R
\name{validation_function}
\alias{validation_function}
\title{Validation Function for Penalized Regression Models}
\usage{
validation_function(input, output, B0, PostB)
}
\arguments{
\item{input}{Dataframe or matrix. The covariates (X) from the dataset used to fit the model.}

\item{output}{Vector. The outcome variable (Y) from the dataset.}

\item{B0}{Dataframe. The true regression coefficients (\code{metric} and \code{value}) for each predictor.}

\item{PostB}{Dataframe. A summary of the posterior regression coefficients (\code{metric} and \code{value}) from the penalized regression model.}
}
\value{
A list containing:
\itemize{
\item \code{COEF}: A dataframe of the re-estimated significant coefficients from the refitted model.
\item \code{MSE}: The Mean Squared Error between the true and estimated regression coefficients.
\item \code{SENS}: The sensitivity, or the proportion of true non-zero coefficients correctly identified as non-zero in the refitted model.
\item \code{SPEC}: The specificity, or the proportion of true zero coefficients correctly identified as zero in the refitted model.
}
}
\description{
This function validates the performance of a penalized regression model by calculating
the Mean Squared Error (MSE), sensitivity (SENS), and specificity (SPEC). It compares the
true regression coefficients (B0) to the posterior coefficients from a fitted model (PostB),
then refits the model using the significant coefficients and evaluates performance metrics based on this refit.
}
\details{
This function takes the true coefficients (B0) and compares them with the posterior estimates (PostB) from the penalized regression model.
It calculates the MSE between the true and estimated coefficients and refits a new model using only the significant coefficients
identified. From this refitted model, sensitivity (the proportion of correctly identified non-zero coefficients)
and specificity (the proportion of correctly identified zero coefficients) are calculated. The function returns the re-estimated
coefficients, sensitivity, and specificity.
}
\examples{
# Example usage:
input <- matrix(rnorm(1000), nrow = 100, ncol = 10)
output <- rbinom(100, 1, 0.5)
B0 <- data.frame(metric = paste0("X", 1:10), value = rnorm(10))
PostB <- data.frame(metric = paste0("X", 1:10), value = rnorm(10))
results <- validation_function(input, output, B0, PostB)
print(results)

}
